teacher-tester-system/requirements.txt
pydantic>=1.9.0
pyyaml>=6.0
pandas>=1.3.0
pytest>=7.0.0